+++ CHANGELOG.md
@@ -25,6 +25,7 @@
   - Support for CSS nesting
 - Upgrade development Node.js LTS to v20.17.0 ([#591](https://github.com/marp-team/marp-cli/pull/591))
 - Upgrade dependent packages to the latest versions ([#591](https://github.com/marp-team/marp-cli/pull/591), [#598](https://github.com/marp-team/marp-cli/pull/598))
+- Migrate ESLint config to flat config ([#606](https://github.com/marp-team/marp-cli/pull/606))
 
 ### Fixed